Move SharedDomain to lib root
