+++ src/core.rs
@@ -1,11 +1,8 @@
 use std::any::Any;
-use std::collections::BTreeSet;
 use std::ptr::{addr_of, NonNull};
 use std::rc::Rc;
 use std::sync::atomic::{AtomicUsize, Ordering::SeqCst};
-use std::sync::{Arc, Mutex};
-
-use crate::stack::SharedStack;
+use std::sync::Arc;
 
 // -------------------------------------
 
@@ -63,96 +60,6 @@ macro_rules! deref_impl {
 deref_impl!(<D: Domain> Domain for &D);
 deref_impl!(<D: Domain> Domain for Rc<D>);
 deref_impl!(<D: Domain> Domain for Arc<D>);
-
-// -------------------------------------
-
-/// Shared, multithreaded domain
-#[derive(Debug)]
-pub struct SharedDomain {
-    hzrd: SharedStack<HzrdPtr>,
-    retired: Mutex<Vec<RetiredPtr>>,
-}
-
-impl SharedDomain {
-    /**
-    Construct a new, clean shared domain
-
-    # Example
-    ```
-    # use hzrd::core::SharedDomain;
-    let domain = SharedDomain::new();
-    ```
-    */
-    pub const fn new() -> Self {
-        Self {
-            hzrd: SharedStack::new(),
-            retired: Mutex::new(Vec::new()),
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn number_of_hzrd_ptrs(&self) -> usize {
-        self.hzrd.iter().count()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn number_of_retired_ptrs(&self) -> usize {
-        self.retired.lock().unwrap().len()
-    }
-}
-
-unsafe impl Domain for SharedDomain {
-    fn hzrd_ptr(&self) -> &HzrdPtr {
-        // Important to only grab shared references to the HzrdPtr's
-        // as others may be looking at them
-        match self.hzrd.iter().find_map(|node| node.try_acquire()) {
-            Some(hzrd_ptr) => hzrd_ptr,
-            None => self.hzrd.push(HzrdPtr::new()),
-        }
-    }
-
-    fn just_retire(&self, ret_ptr: RetiredPtr) {
-        self.retired.lock().unwrap().push(ret_ptr);
-    }
-
-    fn reclaim(&self) {
-        // Try to aquire lock, exit if it is taken
-        let Ok(mut retired_ptrs) = self.retired.try_lock() else {
-            return;
-        };
-
-        // Check if it's empty, no need to move forward otherwise
-        if retired_ptrs.is_empty() {
-            return;
-        }
-
-        let hzrd_ptrs: BTreeSet<_> = self.hzrd.iter().map(HzrdPtr::get).collect();
-        retired_ptrs.retain(|p| hzrd_ptrs.contains(&p.addr()));
-    }
-
-    // -------------------------------------
-
-    // Override this for (hopefully) improved performance
-    fn retire(&self, ret_ptr: RetiredPtr) {
-        // Grab the lock to retired pointers
-        let mut retired_ptrs = self.retired.lock().unwrap();
-
-        // And retire the given pointer
-        retired_ptrs.push(ret_ptr);
-
-        // Check if it's empty, no need to move forward otherwise
-        if retired_ptrs.is_empty() {
-            return;
-        }
-
-        let hzrd_ptrs: BTreeSet<_> = self.hzrd.iter().map(HzrdPtr::get).collect();
-        retired_ptrs.retain(|p| hzrd_ptrs.contains(&p.addr()));
-    }
-}
-
-// -------------------------------------
-
-// TODO: Introduce LocalDomain (`Send` but not `Sync`, use UnsafeCell + LinkedList & Vec)
 
 // -------------------------------------
 
@@ -290,48 +197,8 @@ mod tests {
         unsafe { hzrd_ptr.protect(&mut value) };
     }
 
-    fn new_value<T>(value: T) -> NonNull<T> {
-        let boxed = Box::new(value);
-        let raw = Box::into_raw(boxed);
-        unsafe { NonNull::new_unchecked(raw) }
-    }
-
     #[test]
-    fn retirement() {
-        let value = new_value([1, 2, 3]);
-
-        let domain = SharedDomain::new();
-
-        let hzrd_ptr = domain.hzrd_ptr();
-        unsafe { hzrd_ptr.protect(value.as_ptr()) };
-
-        domain.retire(unsafe { RetiredPtr::new(value) });
-        assert_eq!(domain.number_of_retired_ptrs(), 1);
-
-        domain.reclaim();
-        assert_eq!(domain.number_of_retired_ptrs(), 1);
-
-        unsafe { hzrd_ptr.reset() };
-
-        domain.reclaim();
-        assert_eq!(domain.number_of_retired_ptrs(), 0);
-    }
-
-    #[test]
-    fn domain() {
-        let ptr = new_value(['a', 'b', 'c', 'd']);
-        let domain = SharedDomain::new();
-
-        let hzrd_ptr = domain.hzrd_ptr();
-        unsafe { hzrd_ptr.protect(ptr.as_ptr()) };
-        let set: BTreeSet<_> = domain.hzrd.iter().map(HzrdPtr::get).collect();
-        assert!(set.contains(&(ptr.as_ptr() as usize)));
-
-        domain.retire(unsafe { RetiredPtr::new(ptr) });
-    }
-
-    #[test]
-    fn deep_leak() {
+    fn retired_ptr() {
         let object = vec![String::from("Hello"), String::from("World")];
         let ptr = NonNull::from(Box::leak(Box::new(object)));
